4. PostCommentController로 전환
